+++ src/App.tsx
@@ -1,15 +1,12 @@
 import styled from '@emotion/native';
 import { DefaultTheme, NavigationContainer } from '@react-navigation/native';
-import { useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 import { Text, View } from 'react-native';
 import { useInitializeMsw } from '~hooks/useInitializeMsw';
 import { BottomTabNavigator } from '~navigation/BottomTabNavigator';
 import { AppProviders } from '~providers/AppProviders';
 import { lightTheme } from '~styles/theme';
 import StoryBookUI from '../.storybook';
-import Toast, { BaseToast, BaseToastProps, ErrorToast, ToastConfigParams } from 'react-native-toast-message';
-import { FormErrorToast } from '~components/Common/FormErrorToast';
-import { Animated } from 'react-native';
 
 const navTheme = {
   ...DefaultTheme,
@@ -19,52 +16,10 @@ const navTheme = {
   },
 };
 
-const toastConfig = {
-  success: (props: BaseToastProps) => (
-    <BaseToast
-      {...props}
-      style={{ borderLeftColor: '#783D16' }}
-      contentContainerStyle={{ paddingHorizontal: 15 }}
-      text1Style={{
-        fontFamily: 'SUIT-Bold',
-        fontSize: 13,
-      }}
-    />
-  ),
-  error: (props: BaseToastProps) => (
-    <ErrorToast
-      {...props}
-      style={{ borderLeftColor: 'red' }}
-      text1Style={{
-        fontFamily: 'SUIT-Bold',
-        fontSize: 13,
-        color: 'red',
-      }}
-      text2Style={{
-        fontFamily: 'SUIT-Bold',
-        fontSize: 11,
-      }}
-    />
-  ),
-  formError: ({ text1, props }: ToastConfigParams<{ position: { top: number; left: number } }>) => (
-    <View
-      style={{
-        position: 'absolute',
-        top: props?.position?.top || 0,
-        left: props?.position?.left || 0,
-        transform: [{ translateX: '-50%' }],
-      }}
-    >
-      <FormErrorToast message={text1 || ''} />
-    </View>
-  ),
-};
-
 const MainApp = () => (
   <AppProviders>
     <NavigationContainer theme={navTheme}>
       <BottomTabNavigator />
-      <Toast config={toastConfig} />
     </NavigationContainer>
   </AppProviders>
 );
