+++ src/providers/AppProviders.tsx
@@ -1,11 +1,14 @@
 import { PropsWithChildren } from 'react';
 import { EmotionProvider } from '~providers/EmotionProvider';
 import { TanstackQueryProvider } from '~providers/QueryClientProvider';
+import { ToastProvider } from '~providers/ToastProvider';
 
 export const AppProviders = ({ children }: PropsWithChildren) => {
   return (
     <EmotionProvider>
-      <TanstackQueryProvider>{children}</TanstackQueryProvider>
+      <TanstackQueryProvider>
+        <ToastProvider>{children}</ToastProvider>
+      </TanstackQueryProvider>
     </EmotionProvider>
   );
 };
